refactor: :art: add navbar ui component
