+++ src/apps/App.Web.UI/App.Web.UI.csproj
@@ -6,4 +6,8 @@
     <ImplicitUsings>enable</ImplicitUsings>
   </PropertyGroup>
 
+  <ItemGroup>
+    <Folder Include="Utilities\CustomViewComponent\NavBar\" />
+  </ItemGroup>
+
 </Project>
