+++ src/apps/App.Web.UI/Utilities/CustomViewComponent/NavBar/MessagesDropdownViewComponent.cs
