+++ src/apps/App.Web.UI/Utilities/CustomViewComponent/NavBar/NotificationsDropdownViewComponent.cs
