+++ src/apps/App.Web.UI/Utilities/CustomViewComponent/NavBar/UserMenuViewComponent.cs
